requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
filelock==3.10.0
Flask==2.3.2
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.6
Jinja2==3.1.2
keras==2.12.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0rc1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
torch==2.0.0
torchvision==0.15.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.6
wrapt==1.14.1
